requirements.txt
python-dotenv
google-generativeai
tqdm
pathspec


